requirements.txt
asgiref==3.2.10
Django==3.1.1
django-extensions==3.0.9
djangorestframework==3.12.1
pytz==2020.1
sqlparse==0.3.1
